requirements.txt
requests
feedparser
PyYAML
python-dateutil
scikit-learn
numpy
google-genai
